refactor: Add a reset icon and Modify import-image path
