+++ src/components/common/TextField/index.jsx
@@ -3,7 +3,7 @@ import classNames from 'classnames/bind';
 import Comment from '@/api/comments';
 import IconButton from '@/components/common/button/IconButton';
 import BaseButton from '@/components/common/button/BaseButton';
-import { ICON } from '@/constants';
+import { ICON } from '@/constants/importImage';
 import styles from './TextField.module.scss';
 
 const cx = classNames.bind(styles);
@@ -43,7 +43,7 @@ const TextField = ({ cardId, columnId, dashboardId, name, ...props }) => {
           },
         })}
         {...props}
-        maxLength={255}
+        maxLength={245}
         className={cx('comment-textarea')}
       ></textarea>
       <div className={cx('comment-footer')}>
